Refactor and expand testing for problem solutions
The primary change in this commit includes the addition of more test cases and refactoring of test file structure in "global-problem-solutions.test.ts". Testing for global problem solutions has now been expanded with more test cases, ensuring a more comprehensive testing suite. Some test cases have also been ported over from other test files for better organization of related tests. Additionally, unnecessary code from "seed.test.ts" has been removed and new "wishing-wells.test.ts" file has been created to segregate specific test scenarios.
